+++ example/groq_transcribe_example.dart
@@ -4,14 +4,14 @@ void main(List<String> arguments) async {
   final groq = Groq('YOUR_API_KEY');
 
   //Checking model availability
-  if (!await groq.canUseModel(llama3_8b)) {
+  if (!await groq.canUseModel(GroqModels.llama3_8b)) {
     print('Cannot use model');
     return;
   }
 
   //Sending an audio file for transcription
   final (response, usage, rateLimitInfo) = await groq.transcribeAudio(
-      audioFileUrl: 'res/audio.m4a', modelId: whisper_large_v3);
+      audioFileUrl: 'res/audio.m4a', modelId: GroqModels.whisper_large_v3);
 
   //Printing some received information
   print('Transcription: ${response.text}');